Modulacion de escenas padre competitiva y cooperativa
