+++ src/js/scenes/CooperativeScene.js
@@ -3,28 +3,17 @@ import {Player_I} from "../objects/Player_I.js";
 import {KeyboardProcessor} from "../util/InputProcessors/KeyboardProcessor.js";
 import {SweepVerticalTransitionIn, SweepVerticalTransitionOut} from "../util/cameraEffects.js";
 import {Door} from "../objects/Door.js";
+import {GameStage} from "./GameStage.js";
 
-export class CooperativeScene extends Phaser.Scene {
+export class CooperativeScene extends GameStage {
 
 
     constructor(sceneKey, nextLevelKey, timerTime) {
-        super(sceneKey);
-        this.nextLevelKey = nextLevelKey
-        this.players = [];
-        this.bump = false;
-        this.scores = [];
-        this.door = null;
-        this.music = null;
+        super(sceneKey, nextLevelKey, timerTime,960);
+        
         this.backgroundMusicKey = 'coopStageMusic';
 
-        this.timer = new Timer(this, timerTime)
-        this.timer.onComplete(() => {
-            console.log(
-                this.taskManager.getPlayerWithMoreTasksCompleted()
-            );
-            this.timeOver()
-            this.timer.pauseTimer();
-        })
+
     }
 
     init() {
@@ -32,47 +21,9 @@ export class CooperativeScene extends Phaser.Scene {
     }
 
     create(data,tilemapKey) {
-        this.loadBackgroundMusic()
-        this.playBackgroundMusic()
+        super.create(data,tilemapKey)
 
-        this.setCanvasWidth(960)
-
-        let floor = this.createFloor(tilemapKey);
-
-        this.door = new Door(this, 0, 0,  this.timer)
-
-        this.players[0]  = new Player_I(this, 0, 0, "dude");
-        // data.input1.setPlayer(player1)
-        // player1.setPlayerInput(data.input1);
-        this.players[0].setPlayerInput(new KeyboardProcessor(this, this.players[0] , 'W', 0, 'A', 'D', 'S', 'F'));
-
-        this.players[1] = new Player_I(this, 0, 0, "dude2");
-        this.players[1].setPlayerInput(new KeyboardProcessor(this, this.players[1] , 'U', 0, 'H', 'K', 'J', 'L'));
-
-        this.setPlayersData(data)
-        this.disableAllPlayersMovement()
-
-        this.scores[0] = this.add.text(75, 32, "Player 1: " + this.players[0].points, {fontFamily: 'ink-free-normal'}).setOrigin(.5, .5);
-        this.scores[1] = this.add.text(this.game.canvas.width-75, 32, "Player 2: " + this.players[1].points, {fontFamily: 'ink-free-normal'}).setOrigin(.5, .5);
-
-        this.timer.startTimer();
-        this.timer.pauseTimer();
-
-        this.loadTransition = new SweepVerticalTransitionOut(this);
-        this.loadTransition.addToScene()
-        this.loadTransition.playTransition(() => {
-
-                this.timer.resumeTimer();
-                this.enableAllPlayersMovement()
-            }, 500, 500
-        )
-
-        this.timerText = this.add.text(this.game.canvas.width * 0.5, 40, 'test', {
-            fontFamily: 'ink-free-normal',
-            fontSize: '40px'
-        }).setOrigin(0.5, 0.5);
-
-
+        /// only coop
         let timerTween = this.tweens.add({
             targets: this.timerText,
             paused: true,
@@ -83,64 +34,27 @@ export class CooperativeScene extends Phaser.Scene {
             yoyo: true,
             repeat: 0,            // -1: infinity
         });
-
+        this.door = new Door(this, 0, 0,  this.timer)
         this.taskManager.setOnTaskCompletedTween(timerTween)
 
         this.physics.add.collider(this.players[0], this.door, () => this.door.playerEntered(this.players[0]))
         this.physics.add.collider(this.players[1], this.door, () => this.door.playerEntered(this.players[1]))
-        //***** between players
-        this.physics.add.collider(this.players[0], this.players[1], function () {
-            this.bump = true;
-        });
+        
 
-        this.addStageFloorCollisions( floor);
         //**** players and platforms
         // this.setPlatformsColliders();
     }
 
     update() {
-        this.players[0].update(this.bump, this.players[1]);
-        this.players[1].update(this.bump, this.players[0]);
-        this.bump = false;
-
-        this.timerText.setText(this.timer.getRemainingSeconds(true));
+        super.update();
         this.UpdatePlatforms();
     }
 
-    setPlayersData(data){
-        if(Object.keys(data).length!==0){  // if object has information
-            for (let i = 0; i < this.players.length; i++) {
-                this.players[i].points=data.playerPoints[i];
-            }
-        }
-
-
-    }
-    setPlayerPosition(playerIndex,x,y){
-        this.players[playerIndex].setPosition(x,y);
-    }
-
+    
     setDoorPosition(x,y){
         this.door.setPosition(x,y);
     }
-
-    setCanvasWidth(width) {
-        this.game.canvas.width = width;
-        this.physics.world.setBounds(0, 0, this.game.canvas.width, this.game.canvas.height)
-    }
-
-    createFloor(tilemapKey) {
-        //*************** tilemap
-        let map = this.make.tilemap({key: tilemapKey});
-        let tileset = map.addTilesetImage('Tileset', 'tileset');
-
-        ///************** floor
-        let floor = map.createLayer('Level', tileset);
-        map.createLayer('Background', tileset);
-        floor.setCollisionByProperty({collides: true});
-        floor.depth=2
-        return floor;
-    }
+    
 
     updatePoints(context, playerIndex, points) {
         if ((context.players[playerIndex].points + points) <= 0) {
@@ -161,7 +75,8 @@ export class CooperativeScene extends Phaser.Scene {
         });
         textTween.play()
     }
-    timeOver() {
+
+    onTimeOverPrimitive() {
 
         let playerIndexWithMoreTasksCompleted = this.taskManager.getPlayerWithMoreTasksCompleted();
         this.timerOverUpdatePoints(this, playerIndexWithMoreTasksCompleted, 100)
@@ -179,6 +94,7 @@ export class CooperativeScene extends Phaser.Scene {
         timeOverTimer.startTimer()
 
     }
+    
     timerOverUpdatePoints(context, playerIndex, points) {
         if ((this.players[playerIndex].points + points) <= 0) {
             this.players[playerIndex].points = 0;
@@ -234,21 +150,12 @@ export class CooperativeScene extends Phaser.Scene {
             this.players[i].enableMovement()
         }
     }
-
-    disableAllPlayersMovement() {
-        for (let i = 0; i < this.players.length; i++) {
-            this.players[i].disableMovement()
-        }
-    }
+    
 
     addPointsToPlayer(playerIndex, points) {
         this.players[0].puntos += points;
     }
 
-    startNextLevel() {
-        this.music.stop()
-        this.scene.start(this.nextLevelKey, {playerPoints: [this.players[0].points, this.players[1].points]})
-    }
 
     setPlatformsColliders() {
 
@@ -276,15 +183,4 @@ export class CooperativeScene extends Phaser.Scene {
     }
 
 
-    playBackgroundMusic() {
-        this.music.play();
-    }
-
-    loadBackgroundMusic() {
-        this.music = this.sound.add(this.backgroundMusicKey, {volume: 0.18});
-    }
-
-    stopBackgroundMusic() {
-        this.music.stop()
-    }
 }
